Use assertj fluent style in flowable-bpmn-converter module. This is part 5 of 5 parts with changes suggested by Filip.

+++ modules/flowable-bpmn-converter/src/test/java/org/flowable/editor/language/xml/ValuedDataObjectConverterTest.java
@@ -113,7 +113,7 @@ public class ValuedDataObjectConverterTest extends AbstractConverterTest {
         assertThat(dataObj.getName()).isEqualTo("LongTest");
         assertThat(dataObj.getItemSubjectRef().getStructureRef()).isEqualTo("xsd:long");
         assertThat(dataObj.getValue()).isInstanceOf(Long.class);
-        assertThat(dataObj.getValue()).isEqualTo((long) -123456);
+        assertThat(dataObj.getValue()).isEqualTo(-123456L);
         assertThat(dataObj.getExtensionElements()).hasSize(1);
         List<ExtensionElement> testValues = dataObj.getExtensionElements().get("testvalue");
         assertThat(testValues).isNotNull();
